requirements.txt
Flask==2.3.0
spacy==3.7.2
pdfplumber==0.9.0
nltk==3.8.1
numpy==1.26.4
python-docx==1.1.0
Werkzeug>=2.3.0
